Change model to istupakov/parakeet-tdt-0.6b-v2-onnx and default to fp32

+++ src/hub.js
@@ -395,8 +395,8 @@ export async function getParakeetModel(repoIdOrModelKey, options = {}) {
   const defaultPreprocessor = modelConfig?.preprocessor || 'nemo128';
 
   const {
-    encoderQuant = 'int8',
-    decoderQuant = 'int8',
+    encoderQuant = 'fp32',
+    decoderQuant = 'fp32',
     preprocessorBackend = 'js',
     progress,
   } = options;
